+++ apps/api/services/user_services.py
@@ -1,17 +1,18 @@
 from django.contrib.auth.hashers import make_password
+from django.db.models import QuerySet
 
 from apps.core.models import User
 
 from apps.core.services.random_generate_service import generate_random_number_service
 
 
-def create_user_by_data_service(data: object) -> object:
+def create_user_by_data_service(data: dict) -> QuerySet[User]:
     """
     Функция создает нового пользователя по данным из request
     :param data: request.data
     :return: user
     """
-    user: object = User.objects.create(
+    user: QuerySet[User] = User.objects.create(
         first_name=data['first_name'],
         last_name=data['last_name'],
         email=data['email'],
@@ -20,7 +21,7 @@ def create_user_by_data_service(data: object) -> object:
     return user
 
 
-def get_user_service(user_id: int) -> object:
+def get_user_service(user_id: int) -> QuerySet[User]:
     """
     Функция возвращает user по id
     :param user_id: user id
@@ -29,7 +30,7 @@ def get_user_service(user_id: int) -> object:
     return User.objects.get(pk=user_id)
 
 
-def get_user_by_slug_service(slug: str) -> object:
+def get_user_by_slug_service(slug: str) -> QuerySet[User]:
     """
     Функция возвращает user по slug
     :param slug: slug
@@ -38,7 +39,7 @@ def get_user_by_slug_service(slug: str) -> object:
     return User.objects.get(slug=slug)
 
 
-def verify_user_service(user: object) -> object:
+def verify_user_service(user: QuerySet[User]) -> QuerySet[User]:
     """
     Функция верифицирует пользователя
     :param user: user
@@ -50,19 +51,19 @@ def verify_user_service(user: object) -> object:
     return user
 
 
-def generate_new_verify_code_for_user_service(user_id: int) -> object:
+def generate_new_verify_code_for_user_service(user_id: int) -> QuerySet[User]:
     """
     Функция генерирует новый код подтверждения для user
     :param user_id: user id
     :return: user
     """
-    user: object = get_user_service(user_id)
+    user: QuerySet[User] = get_user_service(user_id)
     user.verify_code = generate_random_number_service(6)
     user.save()
     return user
 
 
-def verify_user_by_code_service(user: object, verify_code: str) -> bool:
+def verify_user_by_code_service(user: QuerySet[User], verify_code: str) -> bool:
     """
     Функция проверяет код и возвращает результат проверки True или False соответственно
     :param user: user